Added Shortid Module & Main App Unit Tests
I needed to use shortid because I was having the potential risk of
generating duplicates of the my short url. As it was just a random
number generator giving a number b/w 1 & 10000, it was difficult to
ensure I had no collisions.

Main app unit tests to ensure the server behaves as expected.